rename template
