+++ tests/test_core.py
@@ -1,4 +1,4 @@
-from python_project_template.core import app
+from er_aws_msk.core import app
 
 
 def test_app() -> None:
